+++ gulp-jsbeautifier/examples/gulpfile.js
@@ -7,6 +7,11 @@ gulp.task('verify-js', function() {
     .pipe(prettify({config: '.jsbeautifyrc', mode: 'VERIFY_ONLY'}));
 });
 
+gulp.task('verify-js-diff', function() {
+  gulp.src(['./src/foo.js', './src/bar.json'])
+    .pipe(prettify({config: '.jsbeautifyrc', mode: 'VERIFY_ONLY', showDiff: false}));
+});
+
 gulp.task('format-js', function() {
   gulp.src(['./src/foo.js', './src/bar.json'])
     .pipe(prettify({config: '.jsbeautifyrc', mode: 'VERIFY_AND_WRITE'}))
@@ -15,12 +20,18 @@ gulp.task('format-js', function() {
 
 gulp.task('prettify-html', function() {
   gulp.src('./src/foo.html')
-    .pipe(prettify({indentSize: 2}))
+    .pipe(prettify({indent_size: 2}))
     .pipe(gulp.dest('./build'));
 });
 
 gulp.task('prettify-css', function() {
   gulp.src('./src/foo.css')
-    .pipe(prettify({indentSize: 2}))
+    .pipe(prettify({indent_size: 2}))
+    .pipe(gulp.dest('./build'));
+});
+
+gulp.task('prettify-css-with-just-css-options', function() {
+  gulp.src('./src/foo.css')
+    .pipe(prettify({css: {indent_size: 2, indent_char: '\t'}}))
     .pipe(gulp.dest('./build'));
 });